Setup of passport-jwt for authentication and routes

+++ controllers/api/v1/posts_api.js
@@ -25,22 +25,21 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
         // .id means converting the object id into string
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
-
-            
-        // req.flash('success', 'Post and associated comments deleted!');
             return res.status(200).json({
                 message: "Post and associated comments deleted successfully"
             })
-        // }else{
-            // req.flash('error', 'You cannot delete this post!');
-            // return res.redirect('back');
-        // }
-    }catch(err){
+        }else{
+            return res.status(401).json({
+                message: "You cannot delete this post!"
+            });
+        }
+    }
+    catch(err){
         // req.flash('err', err);
         return res.status(500).json({
             message: "Internal Server Error"
